requirements.txt
django
djangorestframework
django-cors-headers
drf-spectacular
pillow
PyYAML
